+++ examples/resources/nxos_loopback_interface/resource.tf
@@ -2,4 +2,5 @@ resource "nxos_loopback_interface" "example" {
   interface_id = "lo123"
   admin_state  = "down"
   description  = "My Description"
+  vrf_dn       = "sys/inst-default"
 }