+++ examples/resources/nxos_loopback_interface_vrf/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_loopback_interface_vrf.example
-  id = "<interface_id>"
-}