+++ examples/resources/nxos_loopback_interface_vrf/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_loopback_interface_vrf.example "<interface_id>"