+++ internal/provider/data_source_nxos_loopback_interface.go
@@ -57,7 +57,7 @@ func (d *LoopbackInterfaceDataSource) Metadata(_ context.Context, req datasource
 func (d *LoopbackInterfaceDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: helpers.NewResourceDescription("This data source can read a loopback interface.", "l3LbRtdIf", "Layer%203/l3:LbRtdIf/").String,
+		MarkdownDescription: helpers.NewResourceDescription("This data source can read a loopback interface.", "l3LbRtdIf", "Layer%203/l3:LbRtdIf/").AddAdditionalDocs([]string{"nwRtVrfMbr"}, []string{"Routing%20and%20Forwarding/nw:RtVrfMbr/"}).String,
 
 		Attributes: map[string]schema.Attribute{
 			"device": schema.StringAttribute{
@@ -78,6 +78,10 @@ func (d *LoopbackInterfaceDataSource) Schema(ctx context.Context, req datasource
 			},
 			"description": schema.StringAttribute{
 				MarkdownDescription: "Interface description.",
+				Computed:            true,
+			},
+			"vrf_dn": schema.StringAttribute{
+				MarkdownDescription: "DN of VRF. For example: `sys/inst-VRF1`.",
 				Computed:            true,
 			},
 		},
@@ -114,6 +118,7 @@ func (d *LoopbackInterfaceDataSource) Read(ctx context.Context, req datasource.R
 	}
 
 	queries := []func(*nxos.Req){}
+	queries = append(queries, nxos.Query("rsp-subtree", "children"))
 	res, err := device.Client.GetDn(config.getDn(), queries...)
 	if err != nil {
 		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object, got error: %s", err))