+++ internal/provider/resource_nxos_loopback_interface.go
@@ -63,7 +63,7 @@ func (r *LoopbackInterfaceResource) Metadata(ctx context.Context, req resource.M
 func (r *LoopbackInterfaceResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: helpers.NewResourceDescription("This resource can manage a loopback interface.", "l3LbRtdIf", "Layer%203/l3:LbRtdIf/").AddChildren("loopback_interface_vrf").String,
+		MarkdownDescription: helpers.NewResourceDescription("This resource can manage a loopback interface.", "l3LbRtdIf", "Layer%203/l3:LbRtdIf/").AddAdditionalDocs([]string{"nwRtVrfMbr"}, []string{"Routing%20and%20Forwarding/nw:RtVrfMbr/"}).String,
 
 		Attributes: map[string]schema.Attribute{
 			"device": schema.StringAttribute{
@@ -95,6 +95,10 @@ func (r *LoopbackInterfaceResource) Schema(ctx context.Context, req resource.Sch
 			},
 			"description": schema.StringAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("Interface description.").String,
+				Optional:            true,
+			},
+			"vrf_dn": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("DN of VRF. For example: `sys/inst-VRF1`.").String,
 				Optional:            true,
 			},
 		},
@@ -203,6 +207,7 @@ func (r *LoopbackInterfaceResource) Read(ctx context.Context, req resource.ReadR
 
 	if device.Managed {
 		queries := []func(*nxos.Req){nxos.Query("rsp-prop-include", "config-only")}
+		queries = append(queries, nxos.Query("rsp-subtree", "children"))
 		res, err := device.Client.GetDn(state.Dn.ValueString(), queries...)
 		if err != nil {
 			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object, got error: %s", err))